refactor: move livemanifest package to gitops package
